trajopt_sqp: Convert examples into unit tests
