+++ trajopt_optimizers/trajopt_sqp/include/trajopt_sqp/callbacks/cartesian_error_plotter.h
