+++ trajopt_optimizers/trajopt_sqp/include/trajopt_sqp/callbacks/collision_plotter.h
