+++ trajopt_optimizers/trajopt_sqp/include/trajopt_sqp/callbacks/joint_state_plotter.h
