+++ trajopt_optimizers/trajopt_sqp/include/trajopt_sqp/trust_region_sqp_solver.h
@@ -43,7 +43,7 @@ public:
   OSQPEigenSolver::Ptr qp_solver_;
   QPProblem::Ptr qp_problem_;
 
-  bool verbose_{false};
+  bool verbose_{ false };
 
 private:
   ifopt::Problem* nlp_;
